requirements.txt
fastapi
uvicorn[standard]
cryptography
python-jose
python-multipart
SQLAlchemy
passlib[bcrypt]